refactor: app navigation bar 관련 컴포넌트와 훅 _shared/modules/main으로 이동
